requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.10
PyJWT==2.9.0
Flask-JWT-Extended== 4.7.1
bcrypt==4.3.0
